Made each function into its own script
