+++ config/config.yaml
@@ -23,22 +23,22 @@ qc_method:
   - "seurat"
   - "tutorial"
 
-# # List the dimensionality reduction method(s) that will be used.
-# # Each method must correspond EXACTLY to a function in scripts/dim_reduce.py
-# # If this is left empty or is deleted, dimensionality reduction as well as 
-# # everything downstream will not run (ex. clustering).
-# dim_reduce_method:
-#   - "pca"
-#   - "umap"
-#   - "tsne"
+# List the dimensionality reduction method(s) that will be used.
+# Each method must correspond EXACTLY to a function in scripts/dim_reduce.py
+# If this is left empty or is deleted, dimensionality reduction as well as 
+# everything downstream will not run (ex. clustering).
+dim_reduce_method:
+  - "pca"
+  - "umap"
+  - "tsne"
 
 
-# # List the clustering method(s) that will be used.
-# # Each method must correspond EXACTLY to a function in scripts/cluster.py
-# # If this is left empty or is deleted, no clustering will be performed.
-# cluster_method:
-#   - "leiden"
-#   - "louvain"
+# List the clustering method(s) that will be used.
+# Each method must correspond EXACTLY to a function in scripts/cluster.py
+# If this is left empty or is deleted, no clustering will be performed.
+cluster_method:
+  - "leiden"
+  - "louvain"
 
 
 # too-many-cells:
@@ -92,41 +92,41 @@ qc_method:
 #         - "--filter-tresholds \"(100, 1)\""
 #         - "--draw-collection \"PieRing\""
 
-too-many-cells:
+# too-many-cells:
 
-  make-tree-1:    
-    options:    
-      - --filter-thresholds "(200, 1)"
-      - --draw-collection "PieRing"
+#   make-tree-1:    
+#     options:    
+#       - --filter-thresholds "(200, 1)"
+#       - --draw-collection "PieRing"
   
-  make-tree-2:    
-    prior: "make-tree-1"
-    options:
-      - --draw-node-number
-      - --dendrogram-output "tree-node-numbers.svg"
-      - --draw-collection "PieRing"
-      - --min-size 1
-      - --smart-cutoff 4
+#   make-tree-2:    
+#     prior: "make-tree-1"
+#     options:
+#       - --draw-node-number
+#       - --dendrogram-output "tree-node-numbers.svg"
+#       - --draw-collection "PieRing"
+#       - --min-size 1
+#       - --smart-cutoff 4
 
-  make-tree-3:
-    options:    
-      - --filter-thresholds "(100, 1)"
-      - --draw-collection "PieRing"
+#   make-tree-3:
+#     options:    
+#       - --filter-thresholds "(100, 1)"
+#       - --draw-collection "PieRing"
 
 
 # NOTE: Future feature
 # Plot the results of qc, dimensionality reduction, and clustering
-# plot_method:
+plot_method:
 
-#   scatter:    
-#     use_labels: True   
-#     dim_reduce:
-#       - "pca"
-#       - "umap"
-#       - "tsne"    
-#     cluster:
-#       - "leiden"
-#       - "louvain"
+  scatter:    
+    use_labels: True   
+    dim_reduce:
+      - "pca"
+      - "umap"
+      - "tsne"    
+    cluster:
+      - "leiden"
+      - "louvain"
   
 #   tree:
 #     use_labels: True
