+++ workflow/rules/cluster.smk
@@ -6,18 +6,12 @@ rule cluster:
     input:
         f"{config['output_dir']}/dim_reduce/mtx_{{qc_method}}_{{dr_method}}.h5ad"
     output:
-        # c_method must be a name of a function in cluster.py because
-        # it will be used to call that function
+        # c_method must be a name of a script in the cluster directory
         f"{config['output_dir']}/cluster/mtx_{{qc_method}}_{{dr_method}}_{{c_method}}.h5ad"
     conda:
         "../envs/cluster.yaml"
-    shell:
-        (      
-            f"python {{workflow.basedir}}/scripts/cluster.py "
-            f"{{wildcards.c_method}} "      # Call the function in cluster.py  
-            f"--path_in={{input:q}} "
-            f"--path_out={{output:q}} "
-        )
+    script:
+        "../scripts/cluster/{wildcards.c_method}.py"
 
 
 # def get_maketree_params(wildcards):
