--- workflow/scripts/custom/custom_cluster.py
@@ -1,1 +1,0 @@
-# Insert custom clustering functions in this file
