--- workflow/scripts/custom/custom_dim_reduce.py
@@ -1,1 +1,0 @@
-# Insert custom dimensionality reduction functions in this file
