--- workflow/scripts/custom/custom_qc.py
@@ -1,1 +1,0 @@
-# Insert custom qc functions in this file
